+++ apps/studio/src/lib/projects/index.ts
@@ -2,7 +2,7 @@ import { makeAutoObservable } from 'mobx';
 import { nanoid } from 'nanoid';
 import { sendAnalytics } from '../utils';
 import { MainChannels } from '@onlook/types/constants';
-import type { Project } from '/common/models/project';
+import type { Project } from '@onlook/types/projects';
 import type { AppState, ProjectsCache } from '@onlook/types/settings';
 
 export class ProjectsManager {
